+++ src/helpers/getLocalStorageValue.ts
@@ -1,0 +1,10 @@
+import { initialState } from '../constants/initialState';
+
+export const getLocalStorageValue = () => {
+  const todos = localStorage.getItem('todos');
+
+  return {
+    ...initialState,
+    todos: todos ? JSON.parse(todos) : [],
+  };
+};
